feat: added styling, hard coded path, image not showing
