--- src/components/image/image.stories.tsx
@@ -1,18 +1,0 @@
-import { Meta, StoryFn } from "@storybook/react";
-import { BrowserRouter } from "react-router-dom";
-import { Image } from "./image";
-
-export default {
-  title: "components/image",
-  component: Image,
-} as Meta;
-const Template: StoryFn = () => {
-  return (
-    <BrowserRouter>
-      <Image src="path/to/image.jpg" alt="Description of the image" />
-    </BrowserRouter>
-  );
-};
-
-export const Default = Template.bind({});
-Default.args = {};
