--- src/components/image/image.tsx
@@ -1,9 +1,0 @@
-interface ImageProps {
-  src: string;
-  alt: string;
-  width?: number;
-  height?: number;
-}
-export const Image = ({ src, alt }: ImageProps) => {
-  return <img src={src} alt={alt} />;
-};
